Move the Splash folder to Pages and the move the logic from app to splash.tsx
